perf: shrink VoxelEdit from 32 bytes to 8
Found while stress-testing the flood-fill fix on a 250^3 canvas. The
fill itself now behaves -- the stack stays bounded and the whole 15.6M
voxel block recolors in well under a second -- but the undo group it
records cost ~635 MB, because VoxelEdit stored three usize coordinates
and padded to 32 bytes.

MAX_CHUNK_SIZE is 256, so u16 coordinates leave room to spare and take
the struct to 8 bytes. Same fill measured again: 635 MB -> 310 MB. The
remainder is the Vec's growth doubling, which transiently holds both the
old and new allocation, plus the chunk itself.

A test pins the size, since a regression here silently multiplies what a
large fill costs.

+++ src/state/mod.rs
@@ -232,7 +232,7 @@ impl State {
             // Reverse order so coordinates touched more than once in a single
             // gesture land back on their original value.
             for e in group.iter().rev() {
-                self.chunk.set(e.x, e.y, e.z, e.old);
+                self.chunk.set(e.x as usize, e.y as usize, e.z as usize, e.old);
             }
             self.remesh();
             println!("Undo");
@@ -242,7 +242,7 @@ impl State {
     pub(super) fn redo(&mut self) {
         if let Some(group) = self.history.redo() {
             for e in group.iter() {
-                self.chunk.set(e.x, e.y, e.z, e.new);
+                self.chunk.set(e.x as usize, e.y as usize, e.z as usize, e.new);
             }
             self.remesh();
             println!("Redo");
